Improvements: - return entity saved - delete iteration for interests when board is deleted - join interest with searched tweets

+++ src/main/java/com/fmalessio/twitterapi/service/BoardsServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.fmalessio.twitterapi.entity.Board;
+import com.fmalessio.twitterapi.entity.Interest;
 import com.fmalessio.twitterapi.entity.SearchedTweet;
 import com.fmalessio.twitterapi.repository.BoardRepository;
 import com.fmalessio.twitterapi.repository.SearchedTweetRepository;
@@ -17,15 +18,20 @@ public class BoardsServiceImpl implements BoardsService {
 	private BoardRepository boardRepository;
 	private SearchedTweetRepository searchedTweetRepository;
 
+	private InterestsService interestsService;
+
 	@Autowired
-	public BoardsServiceImpl(BoardRepository boardRepo, SearchedTweetRepository searchedTweetRepository) {
+	public BoardsServiceImpl(BoardRepository boardRepo, SearchedTweetRepository searchedTweetRepository,
+			InterestsService interestsService) {
 		this.boardRepository = boardRepo;
 		this.searchedTweetRepository = searchedTweetRepository;
+		this.interestsService = interestsService;
 	}
 
 	@Override
-	public void createBoard(Board board) {
-		boardRepository.saveAndFlush(board);
+	public Board createBoard(Board board) {
+		board = boardRepository.saveAndFlush(board);
+		return board;
 	}
 
 	@Override
@@ -35,7 +41,15 @@ public class BoardsServiceImpl implements BoardsService {
 
 	@Override
 	public void delete(long id) {
-		boardRepository.deleteById(id);
+		Optional<Board> board = boardRepository.findById(id);
+		if (board.isPresent()) {
+			List<Interest> interests = board.get().getInterests();
+			// It's necessary to remove all Quartz Jobs
+			for (Interest interest : interests) {
+				interestsService.delete(interest.getId());
+			}
+			boardRepository.deleteById(id);
+		}
 	}
 
 	@Override
